+++ packages/core/__tests__/initialPage.e2e.ts
@@ -12,3 +12,16 @@ test('Test the initialPage option', async () => {
     await page.waitForSelector('[data-testid="core__page-layer-6"]', { visible: true });
     await page.waitForFunction(() => 'document.querySelector("[data-testid=core__inner-pages]").scrollTop === 7224');
 });
+
+test('Test the initialPage option: different dimensions', async () => {
+    await page.goto('http://localhost:3000/core-initial-page-different-dimensions');
+    await page.setViewport({
+        width: 1920,
+        height: 1080,
+    });
+    await page.evaluate(() => document.querySelector('[data-testid="core__viewer"]')?.scrollIntoView());
+
+    // Wait until the initial page is rendered
+    await page.waitForSelector('[data-testid="core__page-layer-42"]', { visible: true });
+    await page.waitForFunction(() => 'document.querySelector("[data-testid=core__inner-pages]").scrollTop === 49686');
+});